Add update methods to WrappedConfig
Move various update methods and variables from UpdatingConfig to
WrappedConfig so that a WrappedConfig can be manually updated
without constructing a new one.
